#246: Refactor thin-film files by adding core and analysis modules for better maintainability
