+++ optiland/thin_film/__init__.py
@@ -9,7 +9,8 @@ Units: wavelength in µm, thickness in µm (nm helpers), AOI in radians (deg hel
 
 from __future__ import annotations
 
+from .analysis import SpectralAnalyzer
 from .layer import Layer
 from .stack import ThinFilmStack
 
-__all__ = ["Layer", "ThinFilmStack"]
+__all__ = ["Layer", "ThinFilmStack", "SpectralAnalyzer"]
